% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quarantine.R
\name{add_quarantine}
\alias{add_quarantine}
\title{Change connectivity matrix in quarantine}
\usage{
add_quarantine(delta, efficacy)
}
\arguments{
\item{delta}{the connectivity matrix used in \link{lsdsim}}

\item{efficacy}{the relative reduction in outward-going transmission}
}
\description{
This function changes the delta matrix used in \link{lsdsim} to implement
quanrantine, defined as a reduction in outward-going transmission. The
strategy is to produce a complete new delta matrix where all populations are
in quarantine. Only specific rows of this matrix will then be used during the
simulation as needed.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
