% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_delta.R
\name{make_delta}
\alias{make_delta}
\title{Build a delta matrix from geographic coordinates}
\usage{
make_delta(coords, diffusion = 0)
}
\arguments{
\item{coords}{a 2-columns matrix of xy coordinates}

\item{diffusion}{the value of the diffusion to be used, defined as the
fraction of the force of infection going outside each patch towards the
neighbours}
}
\description{
Here coordinates are provided as x and y in a 2-columns matrix
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
