% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make a 2-D grid}
\usage{
make_grid(size)
}
\arguments{
\item{size}{the size of the square grid}
}
\description{
Builds a square, regular grid of pre-specified size
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
