% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_incub.R, R/rlsd_latent.R
\name{rlsd_latent}
\alias{rlsd_latent}
\title{Draw LSD latent period}
\source{
Gubbins S. Using the basic reproduction number to assess the risk of
transmission of lumpy skin disease virus by biting insects. Transbound
Emerg Dis. 2019;66: 1873–1883.
}
\usage{
rlsd_latent(n, sd = 0.05)

rlsd_latent(n, sd = 0.05)
}
\arguments{
\item{n}{the number of values to draw}

\item{sd}{the standard deviation of the lognormal distribution}
}
\description{
This random number generator draws average latent period for LSD using
literature-driven parameters and log-normal distributions. The latent period
is defined as the time interval between infection of a cattle, and this
cattle becoming infectious.

This random number generator draws average latent period for LSD using
literature-driven parameters and log-normal distributions. The latent period
is defined as the time interval between infection of a cattle, and this
cattle becoming infectious. Note that the value returned and associated
variability is a population average, and does not reflect individual-level
variation.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
