% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_param.R
\name{rlsd_param}
\alias{rlsd_param}
\title{Draw simulation parameters for LSD}
\source{
Gubbins S. Using the basic reproduction number to assess the risk of
transmission of lumpy skin disease virus by biting insects. Transbound
Emerg Dis. 2019;66: 1873–1883.
}
\usage{
rlsd_param(n)
}
\arguments{
\item{n}{the number of values to draw}
}
\value{
A data.frame with one set of parameters per row, including: R0, the
latent period, the infectious period, the transmission rate (beta), rate of
incubation (sigma), and the clearance rate (gamma).
}
\description{
This function creates tables of input parameters for LSD transmission using
estimates from Gubbins et al. (2019), based on virus isolation in cattle
blood, assuming transmission via Stomoxys calcitrans. It uses sub-functions
\link{rlsd_latent}, \link{rlsd_infec_period}, and \link{rlsd_R0}. See these functions to
change default values. Note that the value returned and associated
variability refer to a population average, and do not reflect
individual-level variation.
}
\examples{
x <- rlsd_param(1e3)
head(x)
summary(x)
hist(x$R0, main = "Distribution of R0", xlab = "R0", border = "white")
hist(x$latent, main = "Latent period", 
     xlab = "Latent period (days)", 
     border = "white")
hist(x$infec_period, main = "Infectious period", 
     xlab = "Infectious period (days)", 
     border = "white")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
