% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_pasymp.R
\name{rlsd_pasymp}
\alias{rlsd_pasymp}
\title{Draw proportion of asymptomatic for LSD}
\source{
TBC
}
\usage{
rlsd_pasymp(n, shape1 = 100, shape2 = 100)
}
\arguments{
\item{n}{the number of values to draw}

\item{shape1}{the first shape parameter of the beta distribution; defaults to
100}

\item{shape2}{the second shape parameter of the beta distribution; defaults to
100}
}
\description{
This random number generator draws the proportion of asymptomatic cases for
LSD using literature-driven parameters and beta distributions. See \code{\link[=rbeta]{rbeta()}}
for details on means and variances. .
}
\examples{

hist(
  rlsd_pasymp(1e5), 
  main = "Proportion of asymptomatic cases", 
  xlab = "P (asymptomatic)"
)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
