% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_R0.R
\name{rlsd_R0_I}
\alias{rlsd_R0_I}
\alias{rlsd_R0_A}
\title{Draw LSD basic reproduction numbers}
\source{
Sanz-Bernardo B, Haga IR, Wijesiriwardana N, Basu S, Larner W, Diaz
AV, et al. Quantifying and modeling the acquisition and retention of lumpy
skin disease virus by hematophagus insects reveals clinically but not
subclinically affected cattle are promoters of viral transmission and key
targets for control of disease outbreaks. J Virol. 2021;95.
doi:10.1128/JVI.02239-20
}
\usage{
rlsd_R0_I(n, mu = 20, sd = 0.02)

rlsd_R0_A(n, mu = 2, sd = 0.1)
}
\arguments{
\item{n}{the number of values to draw}

\item{mu}{the population average, defaults to 9.7 for symptomatic
individuals, and 1.2 for asymptomatic cases}

\item{sd}{the standard deviation of the lognormal distribution, defaults to
0.1 for symptomatic individuals, and 0.1 for asymptomatic cases}
}
\description{
This random number generator draws average values of the basic reproduction
number (R0) for LSD using literature-driven parameters and log-normal
distributions. \code{rlsd_R0_I()} is for symptomatic individuals, while
\code{rlsd_R0_A} is for asymptomatic cases. R0 is defined as the average number of
secondary infections per infected cattle in a fully susceptible herd. Note
that the value returned and associated variability refer to a population
average, and do not reflect individual-level variation. The values used here
refer to transmission via vector Stomoxys calcitrans (see Table 4 in source
for other vectors), based on virus isolation in cattle blood.
}
\examples{

opar <- par(no.readonly = TRUE)
par(mfrow = c(2, 1))
hist(rlsd_R0_I(1e5), main = "R0 for symptomatic cases", xlab = "R0")
hist(rlsd_R0_A(1e5), main = "R0 for asymptomatic cases", xlab = "R0")
par(opar)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
