% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lsdsim.R
\name{summary.lsdsim}
\alias{summary.lsdsim}
\title{Summary for lsdsim objects}
\usage{
\method{summary}{lsdsim}(object)
}
\arguments{
\item{object}{an \link{lsdsim} object}
}
\description{
This function sums compartments across all populations in a \link{lsdsim} object,
for each time step. It also reports the number of patches in 'naive',
'response' and 'ring' state at every time step.
}
