% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_latent.R
\name{rlsd_latent}
\alias{rlsd_latent}
\title{Draw LSD latent period}
\source{
Gubbins S. Using the basic reproduction number to assess the risk of
transmission of lumpy skin disease virus by biting insects. Transbound
Emerg Dis. 2019;66: 1873–1883.
}
\usage{
rlsd_latent(n, mu = 8.1, sd = 0.05)
}
\arguments{
\item{n}{the number of values to draw}

\item{mu}{the population average, defaults to 8.1 days}

\item{sd}{the standard deviation of the lognormal distribution, defaults to
0.05}
}
\description{
This random number generator draws average latent period for LSD using
literature-driven parameters and log-normal distributions. The latent period
is defined as the time interval between infection of a cattle, and this
cattle becoming infectious. Reported values refer to virus isolation in
cattle blood. Note that the value returned and associated variability is a
population average, and does not reflect individual-level variation.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
