% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsd_R0.R
\name{rlsd_R0}
\alias{rlsd_R0}
\title{Draw LSD basic reproduction numbers}
\source{
Gubbins S. Using the basic reproduction number to assess the risk of
transmission of lumpy skin disease virus by biting insects. Transbound
Emerg Dis. 2019;66: 1873–1883.
}
\usage{
rlsd_R0(n, mu = 9.7, sd = 0.05)
}
\arguments{
\item{n}{the number of values to draw}

\item{mu}{the population average, defaults to 9.7}

\item{sd}{the standard deviation of the lognormal distribution, defaults to
0.1}
}
\description{
This random number generator draws average values of the basic reproduction
number (R0) for LSD using literature-driven parameters and log-normal
distributions. R0 is defined as the average number of secondary infections
per infected cattle in a fully susceptible herd. Note that the value returned
and associated variability refer to a population average, and do not reflect
individual-level variation. The values used here refer to transmission via
vector Stomoxys calcitrans (see Table 4 in source for other vectors), based
on virus isolation in cattle blood.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
